Split backend database tests
Move schema-alignment tests into backend-specific child modules so the
shared migration helpers stay small and reusable. Move SQLite permission
model tests into their own module with a shared fixture and parameterized
cascade coverage.

Update the users guide path sample fence to use the repository's plaintext
convention for non-code examples.
